+++ ChangeLog.md
@@ -1,6 +1,6 @@
 2.0.0:
 - Disable focusing video player.
-- Add permanent box-shadow under player, not only when focused.
+- Remove box-shadow under player.
 
 1.3.1:
 - Remove jquery, load faster.